api/pyproject.toml
[project]
name = "userstorywriter-agentic-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.0",
    "openai>=2.15.0",
    "openai-agents>=0.6.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "unicorn>=2.1.4",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "ruff>=0.14.11",
]
